major design change to make dbops great again
