+++ anvio/dbops.py
@@ -9,9 +9,7 @@ import sys
 import time
 import copy
 import random
-import hashlib
 import argparse
-import datetime
 import textwrap
 
 from io import StringIO
@@ -24,12 +22,10 @@ import anvio.db as db
 import anvio.tables as t
 import anvio.fastalib as u
 import anvio.utils as utils
-import anvio.kmers as kmers
 import anvio.terminal as terminal
 import anvio.constants as constants
 import anvio.contigops as contigops
 import anvio.filesnpaths as filesnpaths
-import anvio.genecalling as genecalling
 import anvio.ccollections as ccolections
 import anvio.genomestorage as genomestorage
 import anvio.auxiliarydataops as auxiliarydataops
@@ -37,9 +33,13 @@ import anvio.auxiliarydataops as auxiliarydataops
 from anvio.tables.tableops import Table
 from anvio.drivers import Aligners
 from anvio.errors import ConfigError
-from anvio.drivers.hmmer import HMMer
-from anvio.parsers import parser_modules
 from anvio.sequence import get_list_of_outliers
+from anvio.tables.states import TablesForStates
+from anvio.tables.miscdata import TableForItemAdditionalData
+from anvio.tables.genecalls import TablesForGeneCalls
+from anvio.tables.kmers import KMerTablesForContigsAndSplits
+from anvio.tables.ntpositions import TableForNtPositions
+from anvio.tables.contigsplitinfo import TableForContigsInfo, TableForSplitsInfo
 
 
 __author__ = "Developers of anvi'o (see AUTHORS.txt)"
@@ -1727,7 +1727,7 @@ class ProfileSuperclass(object):
                               prematurely. Just so you know.")
             return
 
-        is_profile_db(self.profile_db_path)
+        utils.is_profile_db(self.profile_db_path)
 
         # we have a contigs db? let's see if it's for real.
         if self.contigs_db_path:
@@ -2158,7 +2158,7 @@ class ProfileDatabase:
 
     def init(self):
         if os.path.exists(self.db_path):
-            is_profile_db(self.db_path)
+            utils.is_profile_db(self.db_path)
             self.db = db.DB(self.db_path, anvio.__profile__version__)
             meta_table = self.db.get_table_as_dict('self')
             self.meta = dict([(k, meta_table[k]['value']) for k in meta_table])
@@ -2242,7 +2242,7 @@ class PanDatabase:
 
     def init(self):
         if os.path.exists(self.db_path):
-            is_pan_db(self.db_path)
+            utils.is_pan_db(self.db_path)
             self.db = db.DB(self.db_path, anvio.__pan__version__)
             meta_table = self.db.get_table_as_dict('self')
             self.meta = dict([(k, meta_table[k]['value']) for k in meta_table])
@@ -2325,7 +2325,7 @@ class ContigsDatabase:
 
     def init(self):
         if os.path.exists(self.db_path):
-            is_contigs_db(self.db_path)
+            utils.is_contigs_db(self.db_path)
             self.db = db.DB(self.db_path, anvio.__contigs__version__)
             meta_table = self.db.get_table_as_dict('self')
             self.meta = dict([(k, meta_table[k]['value']) for k in meta_table])
@@ -2553,8 +2553,8 @@ class ContigsDatabase:
         contigs_kmer_table = KMerTablesForContigsAndSplits('kmer_contigs', k=kmer_size)
         splits_kmer_table = KMerTablesForContigsAndSplits('kmer_splits', k=kmer_size)
         nt_positions_table = TableForNtPositions()
-        contigs_info_table = InfoTableForContigs(split_length)
-        splits_info_table = InfoTableForSplits()
+        contigs_info_table = TableForContigsInfo(split_length)
+        splits_info_table = TableForSplitsInfo()
 
         recovered_split_lengths = []
 
@@ -2701,801 +2701,6 @@ class ContigsDatabase:
 ####################################################################################################
 
 
-class InfoTableForContigs:
-    def __init__(self, split_length):
-        self.db_entries = []
-        self.total_nts = 0
-        self.total_contigs = 0
-        self.split_length = split_length
-
-
-    def append(self, seq_id, sequence, gene_start_stops=None):
-        sequence_length = len(sequence)
-        gc_content = utils.get_GC_content_for_sequence(sequence)
-
-        # how many splits will there be?
-        split_start_stops = utils.get_split_start_stops(sequence_length, self.split_length, gene_start_stops)
-
-        self.total_nts += sequence_length
-        self.total_contigs += 1
-        db_entry = tuple([seq_id, sequence_length, gc_content, len(split_start_stops)])
-        self.db_entries.append(db_entry)
-
-        return (sequence_length, split_start_stops, gc_content)
-
-
-    def store(self, db):
-        if len(self.db_entries):
-            db._exec_many('''INSERT INTO %s VALUES (%s)''' % (t.contigs_info_table_name, (','.join(['?'] * len(self.db_entries[0])))), self.db_entries)
-
-
-class InfoTableForSplits:
-    def __init__(self):
-        self.db_entries = []
-        self.total_splits = 0
-
-
-    def append(self, seq_id, sequence, order, start, end, parent_gc_content, parent):
-        self.total_splits += 1
-        sequence_length = len(sequence)
-        db_entry = tuple([seq_id, order, start, end, sequence_length, utils.get_GC_content_for_sequence(sequence), parent_gc_content, parent])
-        self.db_entries.append(db_entry)
-
-
-    def store(self, db):
-        if len(self.db_entries):
-            db._exec_many('''INSERT INTO %s VALUES (%s)''' % (t.splits_info_table_name, (','.join(['?'] * len(self.db_entries[0])))), self.db_entries)
-
-
-class KMerTablesForContigsAndSplits:
-    def __init__(self, table_name, k=4):
-        self.table_name = table_name
-        self.kmers_class = kmers.KMers(k)
-        self.kmers = sorted(list(self.kmers_class.kmers[k]))
-
-        self.kmer_dict = {}
-        self.db_entries = []
-
-        self.kmers_table_structure = ['contig'] + self.kmers
-        self.kmers_table_types = ['text'] + ['numeric'] * len(self.kmers)
-
-
-    def get_kmer_freq(self, sequence):
-        return self.kmers_class.get_kmer_frequency(sequence, dist_metric_safe=True)
-
-
-    def append(self, seq_id, sequence, kmer_freq=None):
-        if not kmer_freq:
-            kmer_freq = self.kmers_class.get_kmer_frequency(sequence, dist_metric_safe=True)
-
-        db_entry = tuple([seq_id] + [kmer_freq[kmer] for kmer in self.kmers])
-        self.db_entries.append(db_entry)
-
-
-    def store(self, db):
-        db.create_table(self.table_name, self.kmers_table_structure, self.kmers_table_types)
-        db._exec_many('''INSERT INTO %s VALUES (%s)''' % (self.table_name, (','.join(['?'] * len(self.kmers_table_structure)))), self.db_entries)
-
-
-class AdditionalAndOrderDataBaseClass(Table, object):
-    """This is a base class for common operations between order and additional data classes."""
-
-    def __init__(self, args):
-        A = lambda x: args.__dict__[x] if x in args.__dict__ else None
-        self.db_path = A('pan_or_profile_db') or A('profile_db') or A('pan_db')
-        self.just_do_it = A('just_do_it')
-
-        if not self.db_path:
-            raise ConfigError("The AdditionalAndOrderDataBaseClass is inherited with an args object that did not\
-                               contain any database path :/ Even though any of the following would\
-                               have worked: `pan_or_profile_db`, `profile_db`, `pan_db` :(")
-
-        if not self.table_name:
-            raise ConfigError("The AdditionalAndOrderDataBaseClass does not know anything about the table it should\
-                               be working with.")
-
-        is_pan_or_profile_db(self.db_path)
-        self.db_type = utils.get_db_type(self.db_path)
-        self.db_version = utils.get_required_version_for_db(self.db_path)
-
-        database = db.DB(self.db_path, self.db_version)
-        self.additional_data_keys = database.get_single_column_from_table(self.table_name, 'data_key')
-        database.disconnect()
-
-        Table.__init__(self, self.db_path, self.db_version, self.run, self.progress)
-
-
-    def populate_from_file(self, additional_data_file_path, skip_check_names=None):
-
-        if skip_check_names is None and is_blank_profile(self.db_path):
-            # FIXME: this BS is here because blank abvi'o profiles do not know what items they have,
-            #        hence the utils.get_all_item_names_from_the_database function eventually explodes if we
-            #        don't skip check names.
-            skip_check_names = True
-
-        filesnpaths.is_file_tab_delimited(additional_data_file_path)
-
-        data_keys = utils.get_columns_of_TAB_delim_file(additional_data_file_path)
-        data_dict = utils.get_TAB_delimited_file_as_dictionary(additional_data_file_path)
-
-        if not len(data_keys):
-            raise ConfigError("There is something wrong with the additional data file for %s at %s.\
-                               It does not seem to have any additional keys for data :/" \
-                                            % (self.target, additional_data_file_path))
-
-        if self.target == 'layer_orders':
-            OrderDataBaseClass.add(self, data_dict, skip_check_names)
-        else:
-            AdditionalDataBaseClass.add(self, data_dict, data_keys, skip_check_names)
-
-
-    def remove(self, data_keys_list):
-        '''Give this guy a list of key for additional data, and watch their demise.'''
-
-        if not isinstance(data_keys_list, list):
-            raise ConfigError("The remove function in AdditionalDataBaseClass wants you to watch\
-                               yourself before you wreck yourself. In other words, can you please\
-                               make sure the keys you send is of type `list` thankyouverymuch?")
-
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-
-        additional_data_keys = sorted(database.get_single_column_from_table(self.table_name, 'data_key', unique=True))
-
-        if not len(additional_data_keys):
-            self.run.info_single('There is nothing to remove --the %s additional data table is already empty :(' % self.target)
-            database.disconnect()
-
-            return
-
-        missing_keys = [k for k in data_keys_list if k not in additional_data_keys]
-        if len(missing_keys) and not self.just_do_it:
-            database.disconnect()
-            raise ConfigError("The following keys you wanted to remove from the items additional data table are\
-                               not really in the table: '%s'. Anvi'o is confused :/" % (', '.join(missing_keys)))
-
-        if data_keys_list:
-            for key in data_keys_list:
-                if key not in additional_data_keys:
-                    # what the hell, user?
-                    return
-
-                database._exec('''DELETE from %s WHERE data_key="%s"''' % (self.table_name, key))
-
-            self.run.warning("%s data for the following keys removed from the database: '%s'. #SAD." % (self.target, ', '.join(data_keys_list)))
-        else:
-            if not self.just_do_it:
-                raise ConfigError("You did not provide a list of data keys to remove, which means you are about to delete everything in the\
-                                   %s additional data table. Just to be on the safe side, anvi'o is looking for a confirmation. If you\
-                                   try again with the --just-do-it flag flag, anvi'o will put on its business socks, and burn this table\
-                                   and everything in it to the ground." % self.target)
-
-            database._exec('''DELETE from %s''' % (self.table_name))
-
-            self.run.warning("All data from the %s additional data table is removed (ouch)." % self.target)
-
-        database.disconnect()
-
-
-    def export(self, output_file_path):
-        filesnpaths.is_output_file_writable(output_file_path)
-
-        if self.target in ['layers', 'items']:
-            keys, data = AdditionalDataBaseClass.get(self)
-        elif self.target in ['layer_orders']:
-            data = OrderDataBaseClass.get(self, native_form=True)
-            keys = ['data_type', 'data_value']
-        else:
-            raise ConfigError("Your target table '%s' does not make any sense" % self.target)
-
-        if not(len(data)):
-            raise ConfigError("Additional data table for %s is empty. There is nothing to export :/" % self.target)
-
-        utils.store_dict_as_TAB_delimited_file(data, output_file_path, headers=[self.target] + keys)
-
-        self.run.info('Output file for %s' % self.target, output_file_path)
-
-
-    def list_data_keys(self):
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-        additional_data_keys = sorted(database.get_single_column_from_table(self.table_name, 'data_key', unique=True))
-
-        if not len(additional_data_keys):
-            self.run.info_single('There are no additional data for %s in this database :/' % self.target, nl_before=1, nl_after=1, mc='red')
-        else:
-            self.run.warning('', 'AVAILABLE DATA KEYS FOR %s (%d FOUND)' % (self.target.upper(), len(additional_data_keys)), lc='yellow')
-            for data_key in additional_data_keys:
-                rows = database.get_some_rows_from_table_as_dict(self.table_name, 'data_key="%s"' % data_key)
-
-                if self.target == 'layer_orders':
-                    self.run.info_single('%s (%s)' % (data_key, list(rows.values())[0]['data_type']),
-                                         nl_after = 1 if data_key == additional_data_keys[-1] else 0)
-                else:
-                    self.run.info_single('%s (%s, describes %d %s)' % (data_key, list(rows.values())[0]['data_type'], len(rows), self.target),
-                                         nl_after = 1 if data_key == additional_data_keys[-1] else 0)
-
-        database.disconnect()
-
-
-    def data_dict_sanity_check(self, data_dict, data_keys_list=None, treat_data_dict_as=None):
-        data_dict_type = treat_data_dict_as or self.target
-
-        if not isinstance(data_dict, dict):
-            raise ConfigError("Nope. Your data must be of type %s, but it is a %s." % (type(dict()), type(data_dict)))
-
-        if not len(data_dict):
-            raise ConfigError("The data sent for sanity check seems to be empty.")
-
-        if data_keys_list and not isinstance(data_keys_list, list):
-            raise ConfigError("List of keys must be of type `list`. Go away (and come back).")
-
-        # FIXME: we have two controls here. The first one is how we work with order data natively. The second one is how it
-        #        looks like when it is read through the .get() member function of the TableForLayerOrders because rest of
-        #        anvi'o does not know how to work with the native format. This should be fixed by teaching the rest of anvi'o
-        #        how to work with order data dicts in the native form.
-        looks_like_layer_orders = sorted(list(data_dict.values())[0].keys()) == sorted(['data_type', 'data_value']) or \
-                                  sorted(list(data_dict.values())[0].keys()) == sorted(['basic', 'newick'])
-
-        if looks_like_layer_orders and data_dict_type is not 'layer_orders':
-            raise ConfigError("The data you sent here seems to describe an order, but you want anvi'o to treat it\
-                               as additional data for %s. Not cool." % self.target)
-
-        if not looks_like_layer_orders and data_dict_type is 'layer_orders':
-            raise ConfigError("The data that claims to be a layer order data do not seem to be one.")
-
-        if data_keys_list:
-            for item_or_layer_name in data_dict:
-                for key in data_keys_list:
-                    if key not in data_dict[item_or_layer_name]:
-                        raise ConfigError("Your data dictionary fails the sanity check since at least one item in it (i.e., %s) is\
-                                           missing any data for the key '%s'." % (item_or_layer_name, key))
-
-
-class OrderDataBaseClass(AdditionalAndOrderDataBaseClass, object):
-    """Implements a base class to deal with tables that keep order data."""
-
-    def __init__(self, args):
-        AdditionalAndOrderDataBaseClass.__init__(self, args)
-
-
-    def get(self, additional_data_keys=None, additional_data_dict=None, native_form=False):
-        """Will return the layer order data dict.
-
-           If `additional_data_keys` and `additional_data_dict` variables are provided, it will return an order dict
-           with additioanl orders generated from data found in those. An example:
-
-                >>> layers_additional_data_table = TableForLayerAdditionalData(args)
-                >>> layer_orders_table = TableForLayerOrders(args)
-                >>> layer_additional_data_keys, layer_additional_data_dict = layers_additional_data_table.get()
-                >>> layer_orders_data_dict = layer_orders_table.get(layer_additional_data_keys, layer_additional_data_dict
-
-           The `native_form` is tricky. Normaly in this function we turn the data read from the table into
-           the legacy data structure anvi'o has been using before all these were implemented. So, by default,
-           we return the data in that legacy format so everyting else would continue working. But if the
-           programmer is interested in exporting the order data as an output file, we don't want to follow that
-           silly legacy format. So the parameter `native_form` simply returns the data in native form. Export
-           functions use the native form. At some point the rest of the anvi'o codebase should be fixed to
-           work only with the native form, and there is no need to this tyranny. Here is a FIXME for that
-           in case one day someone wants to address that.
-        """
-
-        self.progress.new('Recovering layer order data')
-        self.progress.update('...')
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-        order_data = database.get_table_as_dict(self.table_name)
-        database.disconnect()
-        self.progress.end()
-
-        if native_form:
-            return order_data
-
-        d = {}
-        for order_name in order_data:
-            data_type = order_data[order_name]['data_type']
-            data_value = order_data[order_name]['data_value']
-            if data_type == 'newick':
-                d[order_name] = {'newick': data_value, 'basic': None}
-            elif data_type == 'basic':
-                d[order_name] = {'newick': None, 'basic': data_value}
-            else:
-                raise ConfigError("Something is wrong :( Anvi'o just found an entry in %s table with an unrecognized\
-                                   type of '%s'." % (self.target, data_type))
-
-        if additional_data_keys and additional_data_dict:
-            return self.update_orders_dict_using_additional_data_dict(d, additional_data_keys, additional_data_dict)
-        else:
-            return d
-
-
-    def update_orders_dict_using_additional_data_dict(self, order_data_dict, additional_data_keys, additional_data_dict):
-        if order_data_dict:
-            self.data_dict_sanity_check(order_data_dict, treat_data_dict_as='layer_orders')
-
-        self.data_dict_sanity_check(additional_data_dict, data_keys_list=additional_data_keys, treat_data_dict_as='layers')
-
-        # FIXME: here we need to check whether the two dictionaries are in fact 'compatible' with respect to sample names
-        #        they describe.
-
-        for data_key in additional_data_keys:
-            if '!' in data_key:
-                # we don't order stacked bar charts
-                continue
-
-            layer_name_layer_data_tuples = [(additional_data_dict[layer][data_key], layer) for layer in additional_data_dict]
-            order_data_dict['>> ' + data_key] = {'newick': None, 'basic': ','.join([t[1] for t in sorted(layer_name_layer_data_tuples)])}
-            order_data_dict['>> ' + data_key + ' (reverse)'] = {'newick': None, 'basic': ','.join([t[1] for t in sorted(layer_name_layer_data_tuples, reverse=True)])}
-
-        return order_data_dict
-
-
-    def add(self, data_dict, skip_check_names=False):
-        """
-            The default function to add data into the orders table.
-
-             * `data_dict`: this variable for layer orders is expected to follow this format:
-
-                  d = {
-                          'data_key_01': {'data_type': 'newick',
-                                          'data_value': '(item_or_layer_name_01:0.0370199,(item_or_layer_name_02:0.0227268,item_or_layer_name_01:0.0227268)Int3:0.0370199);'
-                                          },
-                          'data_key_02': {'data_type': 'basic',
-                                          'data_value': 'item_or_layer_name_02,item_or_layer_name_01,item_or_layer_name_03'
-                                          },
-                          (...)
-                  }
-        """
-
-        self.data_dict_sanity_check(data_dict)
-
-        if self.target not in ['layer_orders']:
-            raise ConfigError("You are using an OrderDataBaseClass instance to add %s data into your %s database. This is\
-                               illegal and if you are here, it means someone made a mistake somewhere. If you are a user,\
-                               check your flags to make sure you are targeting the right data table. If you are a programmer,\
-                               you are fired." % (self.target, self.db_type))
-
-        self.run.warning(None, 'New %s data...' % self.target, lc="yellow")
-        data_keys_list = list(data_dict.keys())
-        data_key_types = {}
-        for key in data_keys_list:
-            predicted_key_type = data_dict[key]['data_type']
-
-            data_key_types[key] = predicted_key_type
-            self.run.info('Data key "%s"' % key, 'Type: %s' % (data_key_types[key]), \
-                                            nl_after = 1 if key == data_keys_list[-1] else 0)
-
-        # we be responsible here.
-        keys_already_in_db = [c for c in data_keys_list if c in self.additional_data_keys]
-        if len(keys_already_in_db):
-            if self.just_do_it:
-                self.run.warning('The following keys in your data dict will replace the ones that are already\
-                                  in your %s database: %s.' % (self.db_type, ', '.join(keys_already_in_db)))
-
-                self.remove(keys_already_in_db)
-            else:
-                run.info('Data keys already in the db', ', '.join(keys_already_in_db), nl_before=2, mc='red')
-
-                raise ConfigError("Some of the keys in your new order data appear to be in the database already. If you\
-                                   want to replace those in the database with the ones in your new data use the\
-                                   `--just-do-it` flag.")
-
-        if skip_check_names:
-            self.run.warning("You (or the programmer) asked anvi'o to NOT check the consistency of the names of your %s\
-                              between your additional data and the %s database you are attempting to update. So be it.\
-                              Anvi'o will not check anything, but if things don't look the way you expected them to look,\
-                              you will not blame anvi'o for your poorly prepared data, but choose between yourself or\
-                              Obama." % (self.target, self.db_type))
-        else:
-            TableForLayerOrders.check_names(self, data_dict)
-
-        db_entries = []
-        for item_name in data_dict:
-            db_entries.append(tuple([item_name,
-                                     data_dict[item_name]['data_type'],
-                                     data_dict[item_name]['data_value']]))
-
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-        database._exec_many('''INSERT INTO %s VALUES (?,?,?)''' % self.table_name, db_entries)
-        database.disconnect()
-
-        self.run.info('New order data added to the db for %s' % self.target, '%s.' % (', '.join(data_keys_list)))
-
-
-    def get_layer_names(self, data_dict):
-        layer_names = {}
-
-        for data_key in data_dict:
-            try:
-                if data_dict[data_key]['data_type'] == 'newick':
-                    layer_names[data_key] = utils.get_names_order_from_newick_tree(data_dict[data_key]['data_value'])
-                else:
-                    layer_names[data_key] = [s.strip() for s in data_dict[data_key]['data_value'].split(',')]
-            except:
-                raise ConfigError("Parsing the %s data for %s failed :/ We don't know why, because we are lazy. Please\
-                                   take a loook at your input data and figure out :(" \
-                                                                    % (data_dict[data_key]['data_type'], data_key))
-
-        return layer_names
-
-
-class AdditionalDataBaseClass(AdditionalAndOrderDataBaseClass, object):
-    """Implements additional data ops base class.
-
-       See TableForItemAdditionalData or TableForLayerAdditionalData for usage example.
-
-       See AdditionalAndOrderDataBaseClass for inherited functionality.
-    """
-
-    def __init__(self, args):
-        AdditionalAndOrderDataBaseClass.__init__(self, args)
-
-
-    def get(self):
-        """Will return the additional data keys and the dict."""
-
-        self.progress.new('Recovering additional keys and data for %s' % self.target)
-        self.progress.update('...')
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-        additional_data = database.get_table_as_dict(self.table_name)
-        additional_data_keys = database.get_single_column_from_table(self.table_name, 'data_key', unique=True)
-        additional_data_item_names = database.get_single_column_from_table(self.table_name, 'item_name', unique=True)
-        database.disconnect()
-
-        if not len(additional_data_item_names):
-            self.progress.end()
-            return [], {}
-
-        d = {}
-        for additional_data_item_name in additional_data_item_names:
-            d[additional_data_item_name] = {}
-
-        for entry in additional_data.values():
-            additional_data_item_name = entry['item_name']
-            key = entry['data_key']
-            value = entry['data_value']
-
-            if entry['data_type'] in ['int', 'float']:
-                d[additional_data_item_name][key] = eval(entry['data_type'])(value)
-            else:
-                d[additional_data_item_name][key] = value
-
-        for additional_data_item_name in d:
-            for key in additional_data_keys:
-                if key not in d[additional_data_item_name]:
-                    d[additional_data_item_name][key] = None
-
-        self.progress.end()
-
-        return additional_data_keys, d
-
-
-    def add(self, data_dict, data_keys_list, skip_check_names=False):
-        """Function to add data into the item additional data table.
-
-           * `data_dict`: a dictionary for items or layers additional should follow this format:
-
-                d = {
-                        'item_or_layer_name_01': {'data_key_01': value,
-                                                  'data_key_02': value,
-                                                  'data_key_03': value
-                                                  },
-                        'item_or_layer_name_02': {'data_key_01': value,
-                                                  'data_key_03': value,
-                                                  },
-                        (...)
-                    }
-
-           * `data_keys_list`: is a list of keys one or more of which should appear for each item
-                               in `data_dict`.
-        """
-
-        self.data_dict_sanity_check(data_dict, data_keys_list=data_keys_list)
-
-        if self.target not in ['items', 'layers']:
-            raise ConfigError("You are using an AdditionalDataBaseClass instance to add %s data into your %s database. But\
-                               you know what? You can't do that :/ Someone made a mistake somewhere. If you are a user,\
-                               check your flags to make sure you are targeting the right data table. If you are a programmer,\
-                               you are fired." % (self.target, self.db_type))
-
-        self.run.warning(None, 'New %s additional data...' % self.target, lc="yellow")
-        key_types = {}
-        for key in data_keys_list:
-            if '!' in key:
-                predicted_key_type = "stackedbar"
-            else:
-                type_class = utils.get_predicted_type_of_items_in_a_dict(data_dict, key)
-                predicted_key_type = type_class.__name__ if type_class else None
-
-            key_types[key] = predicted_key_type
-            self.run.info('Data key "%s"' % key, 'Predicted type: %s' % (key_types[key]), \
-                                            nl_after = 1 if key == data_keys_list[-1] else 0)
-
-        # we be responsible here.
-        keys_already_in_db = [c for c in data_keys_list if c in self.additional_data_keys]
-        if len(keys_already_in_db):
-            if self.just_do_it:
-                self.run.warning('The following keys in your data dict will replace the ones that are already\
-                                  in your %s database: %s.' % (self.db_type, ', '.join(keys_already_in_db)))
-
-                self.remove(keys_already_in_db)
-            else:
-                run.info('Data keys already in the db', ', '.join(keys_already_in_db), nl_before=2, mc='red')
-
-                raise ConfigError("Some of the data keys in your new data appear to be in the database already. If you\
-                                   want to replace those in the database with the ones in your new data use the\
-                                   `--just-do-it` flag, and watch anvi'o make an exception just for you and complain\
-                                   about nothin' for this once.")
-
-        if skip_check_names:
-            self.run.warning("You (or the programmer) asked anvi'o to NOT check the consistency of the names of your %s\
-                              between your additional data and the %s database you are attempting to update. So be it.\
-                              Anvi'o will not check anything, but if things don't look the way you expected them to look,\
-                              you will not blame anvi'o for your poorly prepared data, but choose between yourself or\
-                              Obama." % (self.target, self.db_type))
-        else:
-            if self.target == 'layers':
-                TableForLayerAdditionalData.check_names(self, data_dict)
-            elif self.target == 'items':
-                TableForItemAdditionalData.check_names(self, data_dict)
-            else:
-                raise ConfigError("Congratulations, you managed to hit an uncharted are in anvi'o. It is cerrtainly very\
-                                   curious how you got here unless you are trying to implement a new functionality.")
-
-        db_entries = []
-        self.set_next_available_id(self.table_name)
-        for item_name in data_dict:
-            for key in data_keys_list:
-                db_entries.append(tuple([self.next_id(self.table_name),
-                                         item_name,
-                                         key,
-                                         data_dict[item_name][key],
-                                         key_types[key]]))
-
-        database = db.DB(self.db_path, utils.get_required_version_for_db(self.db_path))
-        database._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?)''' % self.table_name, db_entries)
-        database.disconnect()
-
-        self.run.info('New data added to the db for your %s' % self.target, '%s.' % (', '.join(data_keys_list)), nl_after=1)
-
-
-class TableForItemAdditionalData(AdditionalDataBaseClass):
-    """
-       This is the class where we maintain the item additional data table in anvi'o
-       pan and profile databases. Related issue: https://github.com/merenlab/anvio/issues/662.
-    """
-    def __init__(self, args, r=run, p=progress):
-        self.run = r
-        self.progress = p
-
-        A = lambda x: args.__dict__[x] if x in args.__dict__ else None
-        self.table_name = A('table_name') or t.item_additional_data_table_name
-
-        self.target = 'items'
-
-        AdditionalDataBaseClass.__init__(self, args)
-
-
-    def check_names(self, data_dict):
-        """Compares item names found in the data dict to the ones in the db"""
-
-        items_in_db = utils.get_all_item_names_from_the_database(self.db_path)
-        items_in_data = set(data_dict.keys())
-
-        items_in_data_but_not_in_db = items_in_data.difference(items_in_db)
-        if len(items_in_data_but_not_in_db):
-            raise ConfigError("Well. %d of %d item names in your additional data are only in your data (which\
-                               that they are not in the %s database you are working with (which means bad news)).\
-                               Since there is no reason to add additional data for items that do not exist in your\
-                               database, anvi'o will stop you right there. Please fix your data and come again. In\
-                               case you want to see a random item that is only in your data, here is one: %s. Stuff\
-                               in your db looks like this: %s." \
-                                    % (len(items_in_data_but_not_in_db), len(items_in_data), self.db_type, \
-                                       items_in_data_but_not_in_db.pop(), items_in_db.pop()))
-
-        items_in_db_but_not_in_data = items_in_db.difference(items_in_data)
-        if len(items_in_db_but_not_in_data):
-            self.run.warning("Your input contains additional data for only %d of %d total number of items in your %s\
-                              database. Just wanted to make sure you know what's up, but we cool." \
-                                % (len(items_in_db) - len(items_in_db_but_not_in_data), len(items_in_db), self.db_type))
-
-
-class TableForLayerAdditionalData(AdditionalDataBaseClass):
-    """
-       This is the class where we maintain the layer additional data table in anvi'o
-       pan and profile databases.
-
-       Once upon a time there was something called an anvi'o samples database. This
-       is one of two tables that made it irrelevant.
-
-       Related issue: https://github.com/merenlab/anvio/issues/674.
-    """
-
-    def __init__(self, args, r=run, p=progress):
-        self.run = r
-        self.progress = p
-
-        A = lambda x: args.__dict__[x] if x in args.__dict__ else None
-        self.table_name = A('table_name') or t.layer_additional_data_table_name
-
-        self.target = 'layers'
-
-        AdditionalDataBaseClass.__init__(self, args)
-
-
-    def check_names(self, data_dict):
-        """Compares layer names found in the data dict to the ones in the db"""
-
-        layers_in_db = utils.get_all_sample_names_from_the_database(self.db_path)
-        layers_in_data = set(data_dict.keys())
-
-        layers_in_data_but_not_in_db = layers_in_data.difference(layers_in_db)
-        if len(layers_in_data_but_not_in_db):
-            raise ConfigError("Grande problemo. %d of %d layers in your additional data are *only* in your data (which\
-                               means they are not in the %s database you are working with). Since there is no reason to\
-                               add additional data for layers that do not exist in your database, anvi'o refuses to\
-                               continue, and hopes that you will try again. In case you want to see a random layer name\
-                               that is only in your data, here is one: %s. Stuff in your db looks like this: %s." \
-                                    % (len(layers_in_data_but_not_in_db), len(layers_in_data), self.db_type, \
-                                       layers_in_data_but_not_in_db.pop(), layers_in_db.pop()))
-
-        layers_in_db_but_not_in_data = layers_in_db.difference(layers_in_data)
-        if len(layers_in_db_but_not_in_data):
-            self.run.warning("Your input contains additional data for only %d of %d total number of layers in your %s\
-                              database. Just wanted to make sure you know what's up, but we cool." \
-                                % (len(layers_in_db) - len(layers_in_db_but_not_in_data), len(layers_in_db), self.db_type))
-
-
-class TableForLayerOrders(OrderDataBaseClass):
-    """
-       This is the class where we maintain the layer order data table in anvi'o pan and profile
-       databases.
-    """
-
-    def __init__(self, args, r=run, p=progress):
-        self.run = r
-        self.progress = p
-
-        A = lambda x: args.__dict__[x] if x in args.__dict__ else None
-        self.table_name = A('table_name') or t.layer_orders_table_name
-
-        self.allowde_types = ['newick', 'basic']
-        self.target = 'layer_orders'
-
-        OrderDataBaseClass.__init__(self, args)
-
-
-    def check_names(self, data_dict):
-        """Compares layer names found in the data dict to the ones in the db"""
-
-        layers_in_db = sorted(utils.get_all_sample_names_from_the_database(self.db_path))
-        layers_in_data = self.get_layer_names(data_dict)
-
-        for data_key in data_dict:
-            if sorted(layers_in_data[data_key]) != layers_in_db:
-                raise ConfigError("Layer orders data must match the layer names stored in the %s database :/ But at least one of\
-                                   your layer order data, '%s' (a %s order), tells a different story. It has layer names '%s' while\
-                                   the db has the layers '%s' :/" % (self.db_type,
-                                                                     data_key,
-                                                                     data_dict[data_key]['order_type'],
-                                                                     ', '.join(sorted(layers_in_data[data_key])),
-                                                                     ', '.join(layers_in_db)))
-
-
-class MiscDataTableFactory(TableForItemAdditionalData, TableForLayerAdditionalData, TableForLayerOrders):
-    """Gives seamless access to additional data or order tables in pan or profile databases.
-
-       Create an instance with args.target_data_table = [items|layers|layer_orders], and you will be golden.
-    """
-
-    def __init__(self, args, r=run, p=progress):
-        self.run = r
-        self.progress = p
-
-        A = lambda x: args.__dict__[x] if x in args.__dict__ else None
-        target_data_table = A('target_data_table')
-
-        if not target_data_table:
-            raise ConfigError("When creating an instance from the MiscDataTableFactory class, the `args` object\
-                               must contain the `target_data_table` variable.")
-
-        if target_data_table == 'items':
-            TableForItemAdditionalData.__init__(self, args)
-        elif target_data_table == 'layers':
-            TableForLayerAdditionalData.__init__(self, args)
-        elif target_data_table == 'layer_orders':
-            TableForLayerOrders.__init__(self, args)
-        else:
-            raise ConfigError("MiscDataTableFactory does not know about target data tables for '%s' :(\
-                               You can go to the online documentation, or you can try either 'items', 'layers',\
-                               or 'layer_orders'" % target_data_table)
-
-
-class TablesForViews(Table):
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-
-        Table.__init__(self, self.db_path, utils.get_required_version_for_db(db_path), run, progress)
-
-
-    def create_new_view(self, data_dict, table_name, table_structure, table_types, view_name=None, append_mode=False):
-        """Creates a new view table, and adds an entry for it into the 'views' table.
-
-        Entries in 'views' table appear in various places in the interface. However, we also generate
-        view tables to store the type of data we do not wish to display on interfaces, but be able
-        access from various other modules. A good example to this is the item_order recipes. When we
-        profile a sample, we treat every stplit as their own entity with respect to their mean coverage.
-        Although it is great for visualization purposes, it is not useful for item_order purposes since in
-        most cases we wish splits to stay together in item_order output. Hence, we create a mean_coverage_splits
-        table, where each split holds their own coverage, and we create a mean_coverage_contigs table where each
-        split has the coverage of their parent. Clearly the second table is not useful to display. When a table
-        is not added as an entry to the 'views' table, then it only exists in the database for other purposes
-        than displaying it.
-
-        If a new view does not have a 'view_id', it is not added the 'views' table to provide that flexibility.
-        """
-
-        anvio_db = DBClassFactory().get_db_object(self.db_path)
-
-        views_in_db = anvio_db.db.get_table_as_dict(t.views_table_name)
-
-        if not append_mode:
-            if view_name and view_name in views_in_db:
-                raise ConfigError("TablesForViews speaking: Yo yo yo. You already have a view in the db called '%s'.\
-                                    You can't create another one before you get rid of the existing one, because rules."\
-                                                                            % view_name)
-
-            # first create the data table:
-            anvio_db.db.drop_table(table_name)
-
-        try:
-            anvio_db.db.create_table(table_name, table_structure, table_types)
-        except:
-            if not append_mode:
-                raise ConfigError("Table already exists")
-
-        db_entries = [tuple([item] + [data_dict[item][h] for h in table_structure[1:]]) for item in data_dict]
-        anvio_db.db._exec_many('''INSERT INTO %s VALUES (%s)''' % (table_name, ','.join(['?'] * len(table_structure))), db_entries)
-
-        if view_name and view_name not in views_in_db:
-            anvio_db.db._exec('''INSERT INTO %s VALUES (?,?)''' % t.views_table_name, (view_name, table_name))
-
-        anvio_db.disconnect()
-
-
-    def remove(self, view_name, table_names_to_blank=[]):
-        anvio_db = DBClassFactory().get_db_object(self.db_path)
-        anvio_db.db._exec('''DELETE FROM %s WHERE view_id = "%s"''' % (t.views_table_name, view_name))
-        for table_name in table_names_to_blank:
-            if table_name in anvio_db.db.get_table_names():
-                anvio_db.db._exec('''DELETE FROM %s''' % table_name)
-        anvio_db.disconnect()
-
-
-class TableForVariability(Table):
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-        self.run = run
-        self.progress = progress
-
-        Table.__init__(self, self.db_path, utils.get_required_version_for_db(db_path), run=self.run, progress=self.progress)
-
-        self.num_entries = 0
-        self.db_entries = []
-        self.set_next_available_id(t.variable_nts_table_name)
-
-
-    def append(self, profile, quiet=False):
-        db_entry = tuple([self.next_id(t.variable_nts_table_name)] + [profile[h] for h in t.variable_nts_table_structure[1:]])
-        self.db_entries.append(db_entry)
-        self.num_entries += 1
-        if not quiet and self.num_entries % 100 == 0:
-            self.progress.update('Information for %d SNV sites have been added ...' % self.num_entries)
-
-
-    def store(self):
-        profile_db = ProfileDatabase(self.db_path)
-        profile_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)''' % t.variable_nts_table_name, self.db_entries)
-        profile_db.disconnect()
-
-
 class AA_counts(ContigsSuperclass):
     def __init__(self, args, run=run, progress=progress):
         self.args = args
@@ -3622,1073 +2827,11 @@ class AA_counts(ContigsSuperclass):
         return self.counts_dict
 
 
-class TableForAAFrequencies(Table):
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-        self.run = run
-        self.progress = progress
-
-        Table.__init__(self, self.db_path, utils.get_required_version_for_db(db_path), run=self.run, progress=self.progress)
-
-        self.num_entries = 0
-        self.db_entries = []
-        self.set_next_available_id(t.variable_aas_table_name)
-
-
-    def append(self, profile):
-        db_entry = tuple([self.next_id(t.variable_aas_table_name)] + [profile[h] for h in t.variable_aas_table_structure[1:]])
-        self.db_entries.append(db_entry)
-        self.num_entries += 1
-        if self.num_entries % 100 == 0:
-            self.progress.update('Information for %d codons have been added ...' % self.num_entries)
-
-
-    def store(self):
-        profile_db = ProfileDatabase(self.db_path)
-        profile_db.db._exec_many('''INSERT INTO %s VALUES (%s)''' % (t.variable_aas_table_name, ','.join(['?'] * len(t.variable_aas_table_structure))), self.db_entries)
-        profile_db.disconnect()
-
-
-class TablesForGeneCalls(Table):
-    def __init__(self, db_path, contigs_fasta=None, run=run, progress=progress, debug=False):
-        self.run = run
-        self.progress = progress
-        self.db_path = db_path
-        self.contigs_fasta = contigs_fasta
-        self.debug = debug
-
-        is_contigs_db(self.db_path)
-
-        if self.contigs_fasta:
-            filesnpaths.is_file_exists(self.contigs_fasta)
-            filesnpaths.is_file_fasta_formatted(self.contigs_fasta)
-
-
-    def check_gene_calls_dict(self, gene_calls_dict):
-        if not isinstance(gene_calls_dict, type({})):
-            raise ConfigError("Gene calls dict must be a dict instance :/")
-
-        try:
-            [int(g) for g in list(gene_calls_dict.keys())]
-        except ValueError:
-            raise ConfigError("Keys of a gene calls dict must be integers!")
-
-        if False in [x['direction'] in ['f', 'r'] for x in list(gene_calls_dict.values())]:
-            raise ConfigError("The values in 'direction' column can't be anything but 'f' (for forward)\
-                                or 'r' (for reverse). You have other stuff, and it is not cool.")
-
-        if False in [x['stop'] > x['start'] for x in list(gene_calls_dict.values())]:
-            raise ConfigError("For each gene call, the stop position must be bigger than the start position.\
-                                Your gene calls dict does not conform to that. If you have reverse gene calls\
-                                you must use the 'direction' column to declare that.")
-
-        if False in [(x['stop'] - float(x['start'])) % 3.0 == 0 for x in list(gene_calls_dict.values())]:
-            raise ConfigError("Something is wrong with your gene calls. For every gene call, the (stop - start)\
-                                should be multiply of 3. It is not the case for all, which is a deal breaker.")
-
-
-    def use_external_gene_calls_to_populate_genes_in_contigs_table(self, input_file_path, gene_calls_dict=None, ignore_internal_stop_codons=False):
-        """Add genes to the contigs database.
-
-           Either provide an `input_file_path` for external gene calls, or provide an
-           external gene calls dictionary. The format should follow this:
-
-                {
-                  "1": {
-                      "contig": "contig_name",
-                      "start": 20,
-                      "stop": 1544,
-                      "direction": "f",
-                      "partial": 0,
-                      "source": "source_name",
-                      "version": "unknown"
-                  },
-
-                  "2": {
-                    (...)
-                  },
-
-                (...)
-                }
-
-            If you provide a `gene_calls_dict`, they will be APPENDED to the database. So you
-            need to make sure gene caller ids in your dict does not overlap with the ones in
-            the database.
-
-        """
-
-        # by default we assume that this is a pristine run. but if the user sends a dictionary
-        append_to_the_db = False
-
-        gene_calls_found = False
-        # let's do a rigorous check whether the user provided a gene_calls_dict.
-        if (gene_calls_dict is not None and gene_calls_dict is not False):
-            if not isinstance(gene_calls_dict, dict):
-                raise ConfigError("'Use external gene calls' function received a non-empty gene_calls_dict object,\
-                                    but it is of type '%s', and not '%s'" % (type(gene_calls_dict), type({})))
-
-            # congrats, we have a dict.
-            gene_calls_found = True
-
-            if not len(gene_calls_dict):
-                # but it is empty ... silly user.
-                self.run.info_single("'Use external gene calls' function found an empty gene calls dict, returning\
-                                      prematurely and assuming you know what's up. If you don't, stop here and try to\
-                                      identify what decisions you've made might have led you to this weird point your\
-                                      workflow (or 'life', totally up to you and your mood, but anvi'o thinks you've\
-                                      done great so far.", nl_before=1, nl_after=1)
-                return
-
-
-        if (not input_file_path and not gene_calls_found) or (input_file_path and gene_calls_found):
-            raise ConfigError("You must provide either an input file, or an gene calls dict to process external\
-                               gene calls. You called `use_external_gene_calls_to_populate_genes_in_contigs_table`\
-                               with wrong parameters.")
-
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, self.run, self.progress, simple=True)
-
-        # take care of gene calls dict
-        if not gene_calls_found:
-            gene_calls_dict = utils.get_TAB_delimited_file_as_dictionary(input_file_path,
-                                                                         expected_fields=t.genes_in_contigs_table_structure,
-                                                                         only_expected_fields=True,
-                                                                         column_mapping=[int, str, int, int, str, int, str, str])
-
-            if not len(gene_calls_dict):
-                raise ConfigError("You provided an external gene calls file, but it returned zero gene calls. Assuming that\
-                                   this is an error, anvi'o will stop here and complain. If this is not an error and you\
-                                   in fact expected this, the proper way of doing this is to use `--skip-gene-calls` flag,\
-                                   instead of providing an emtpy external gene calls file. You don't agree? You need this\
-                                   for some weird step for you weird pipeline? Let us know, and we will consider changing\
-                                   this.")
-
-            self.run.info("External gene calls", "%d gene calls recovered and will be processed." % len(gene_calls_dict))
-        else:
-            # FIXME: we need to make sure the gene caller ids in the incoming directory is not going to
-            #        overwrite an existing gene call. Something like this would have returned the
-            #        current max, which could be cross-checked with what's in the dict:
-            #
-            #            contigs_db = ContigsDatabase(self.db_path)
-            #            next_id = contigs_db.db.get_max_value_in_column('genes_in_contigs', 'gene_callers_id') + 1
-            #            contigs_db.disconnect()
-            append_to_the_db = True
-
-        # recover amino acid sequences. during this operation we are going to have to read all contig sequences
-        # into the damn memory. anvi'o is doing a pretty bad job with memory management :(
-        amino_acid_sequences = {}
-
-        contig_sequences = {}
-        if self.contigs_fasta:
-            fasta = u.SequenceSource(self.contigs_fasta)
-            while next(fasta):
-                contig_sequences[fasta.id] = {'sequence': fasta.seq}
-            fasta.close()
-        else:
-            class Args: None
-            args = Args()
-            args.contigs_db = self.db_path
-            contigs_db = ContigsSuperclass(args, r=terminal.Run(verbose=False))
-            contigs_db.init_contig_sequences()
-            contig_sequences = contigs_db.contig_sequences
-
-        num_genes_with_internal_stops = 0
-        number_of_impartial_gene_calls = 0
-        for gene_callers_id in gene_calls_dict:
-            gene_call = gene_calls_dict[gene_callers_id]
-            contig_name = gene_call['contig']
-
-            if contig_name not in contig_sequences:
-                # remove the partial contigs database so things don't get screwed later
-                os.remove(self.db_path)
-                raise ConfigError("You are in big trouble :( The contig name '%s' in your external gene callers file\
-                                    does not appear to be in the contigs FASTA file. How did this happen?" % contig_name)
-
-            if gene_call['partial']:
-                amino_acid_sequences[gene_callers_id] = ''
-                number_of_impartial_gene_calls += 1
-                continue
-
-            sequence = contig_sequences[contig_name]['sequence'][gene_call['start']:gene_call['stop']]
-            if gene_call['direction'] == 'r':
-                sequence = utils.rev_comp(sequence)
-
-            amino_acid_sequence = utils.get_DNA_sequence_translated(sequence, gene_callers_id)
-
-            # check if there are any internal stops:
-            if amino_acid_sequence.find('*') > -1:
-                if ignore_internal_stop_codons:
-                    amino_acid_sequence = amino_acid_sequence.replace('*', 'X')
-                    num_genes_with_internal_stops += 1
-                else:
-                    os.remove(self.db_path)
-                    raise ConfigError("Oops. Anvi'o run into an amino acid seqeunce (that corresponds to the gene callers id '%s')\
-                                       which had an internal stop codon :/ This usually indicates that your external gene calls\
-                                       have problems. If you still want to continue, you can ask anvi'o to ignore internal stop\
-                                       codons on your own risk. It will probably look very ugly on your screen, but here is the\
-                                       DNA sequence for that gene in case you don't trust anvi'o (which only would be fair since\
-                                       anvi'o does not trust you either): %s" % (str(gene_callers_id), sequence))
-
-            amino_acid_sequences[gene_callers_id] = amino_acid_sequence
-
-        # populate genes_in_contigs, and gene_amino_acid_sequences table in contigs db.
-        self.populate_genes_in_contigs_table(gene_calls_dict, amino_acid_sequences, append_to_the_db=append_to_the_db)
-
-        if num_genes_with_internal_stops:
-            percent_genes_with_internal_stops = num_genes_with_internal_stops * 100.0 / len(gene_calls_dict)
-            self.run.warning("Please read this carefully: Your external gene calls contained open reading frames with internal\
-                              stop codons, and you asked anvi'o to ignore those. Anvi'o replaced internal stop codons with 'X'\
-                              characters, and stored them in the contigs database that way. %d of your genes, which corresponded\
-                              to %.2f%% of the total %d genes, had internal stop codons. We hope you are happy." % \
-                                        (num_genes_with_internal_stops, percent_genes_with_internal_stops, len(gene_calls_dict)))
-
-        if number_of_impartial_gene_calls:
-            self.run.warning('%d of your %d gene calls were impartial, hence the translated amino acid sequences for those\
-                              were not stored in the database.' % (number_of_impartial_gene_calls, len(gene_calls_dict)))
-
-
-    def call_genes_and_populate_genes_in_contigs_table(self, gene_caller='prodigal'):
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, run, progress, simple=True)
-
-        # get gene calls and amino acid sequences
-        gene_calls_dict, amino_acid_sequences = self.run_gene_caller(gene_caller)
-
-        # make sure the returning gene calls dict is proper
-        self.check_gene_calls_dict(gene_calls_dict)
-
-        # populate genes_in_contigs, and gene_amino_acid_sequences table in contigs db.
-        self.populate_genes_in_contigs_table(gene_calls_dict, amino_acid_sequences)
-
-
-    def run_gene_caller(self, gene_caller='prodigal'):
-        """Runs gene caller, and returns gene_calls_dict, and amino acid sequences."""
-        remove_fasta_after_processing = False
-
-        if not self.contigs_fasta:
-            self.contigs_fasta = self.export_sequences_table_in_db_into_FASTA_file()
-            remove_fasta_after_processing = True
-
-        if self.debug:
-            self.run.info_single('--debug flag is [ON], which means temporary directories generated by\
-                                 this run will not be removed', nl_after=2)
-
-        gene_caller = genecalling.GeneCaller(self.contigs_fasta, gene_caller=gene_caller, debug=self.debug)
-
-        gene_calls_dict, amino_acid_sequences = gene_caller.process()
-
-        if not self.debug and remove_fasta_after_processing:
-            os.remove(self.contigs_fasta)
-
-        return gene_calls_dict, amino_acid_sequences
-
-
-    def populate_genes_in_contigs_table(self, gene_calls_dict, amino_acid_sequences, append_to_the_db=False):
-        contigs_db = db.DB(self.db_path, anvio.__contigs__version__)
-
-        if not append_to_the_db:
-            contigs_db._exec('''DELETE FROM %s''' % (t.genes_in_contigs_table_name))
-            contigs_db._exec('''DELETE FROM %s''' % (t.gene_amino_acid_sequences_table_name))
-        else:
-            # so we are in the append mode. We must remove all the previous entries from genes in contigs
-            # that matches to the incoming sources. otherwhise we may end up with many duplicates in the db.
-            sources = set([v['source'] for v in gene_calls_dict.values()])
-            for source in sources:
-                contigs_db._exec('''DELETE FROM %s WHERE source = "%s"''' % (t.genes_in_contigs_table_name, source))
-
-        self.progress.new('Processing')
-        self.progress.update('Entering %d gene calls into the db ...' % (len(gene_calls_dict)))
-
-        db_entries = [tuple([entry_id] + [gene_calls_dict[entry_id][h] for h in t.genes_in_contigs_table_structure[1:]]) for entry_id in gene_calls_dict]
-        contigs_db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?,?,?)''' % t.genes_in_contigs_table_name, db_entries)
-
-        db_entries = [tuple([entry_id] + [amino_acid_sequences[entry_id]]) for entry_id in gene_calls_dict]
-        contigs_db._exec_many('''INSERT INTO %s VALUES (?,?)''' % t.gene_amino_acid_sequences_table_name, db_entries)
-
-        self.progress.end()
-
-        contigs_db.disconnect()
-
-
-    def populate_genes_in_splits_tables(self):
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, run, progress)
-        self.init_gene_calls_dict()
-
-        genes_in_splits = GenesInSplits()
-        # build a dictionary for fast access to all genes identified within a contig
-        gene_calls_in_contigs_dict = {}
-        for gene_callers_id in self.gene_calls_dict:
-            contig = self.gene_calls_dict[gene_callers_id]['contig']
-            if contig in gene_calls_in_contigs_dict:
-                gene_calls_in_contigs_dict[contig].add(gene_callers_id)
-            else:
-                gene_calls_in_contigs_dict[contig] = set([gene_callers_id])
-
-        contigs_without_any_gene_calls = list(set(self.contigs_info.keys()) - set(gene_calls_in_contigs_dict.keys()))
-        run.info('Contigs with at least one gene call', '%d of %d (%.1f%%)' % (len(gene_calls_in_contigs_dict),
-                                                                               len(self.contigs_info),
-                                                                               len(gene_calls_in_contigs_dict) * 100.0 / len(self.contigs_info)))
-
-        for contig in contigs_without_any_gene_calls:
-            gene_calls_in_contigs_dict[contig] = set([])
-
-        splits_dict = {}
-        for contig in self.contigs_info:
-            for split_name in self.contig_name_to_splits[contig]:
-                start = self.splits_info[split_name]['start']
-                stop = self.splits_info[split_name]['end']
-
-                gene_start_stops = []
-                # here we go through all genes in the contig and identify the all the ones that happen to be in
-                # this particular split to generate summarized info for each split. BUT one important that is done
-                # in the following loop is genes_in_splits.add call, which populates GenesInSplits class.
-                for gene_callers_id in gene_calls_in_contigs_dict[contig]:
-                    if self.gene_calls_dict[gene_callers_id]['stop'] > start and self.gene_calls_dict[gene_callers_id]['start'] < stop:
-                        gene_start_stops.append((self.gene_calls_dict[gene_callers_id]['start'], self.gene_calls_dict[gene_callers_id]['stop']), )
-                        genes_in_splits.add(split_name, start, stop, gene_callers_id, self.gene_calls_dict[gene_callers_id]['start'], self.gene_calls_dict[gene_callers_id]['stop'])
-
-                # here we identify genes that are associated with a split even if one base of the gene spills into
-                # the defined start or stop of a split, which means, split N, will include genes A, B and C in this
-                # scenario:
-                #
-                # contig: (...)------[ gene A ]--------[     gene B    ]----[gene C]---------[    gene D    ]-----(...)
-                #         (...)----------x---------------------------------------x--------------------------------(...)
-                #                        ^ (split N start)                       ^ (split N stop)
-                #                        |                                       |
-                #                        |<-              split N              ->|
-                #
-                # however, when looking at the coding versus non-coding nucleotide ratios in a split, we have to make
-                # sure that only the relevant portion of gene A and gene C is counted:
-                total_coding_nts = 0
-                for gene_start, gene_stop in gene_start_stops:
-                    total_coding_nts += (gene_stop if gene_stop < stop else stop) - (gene_start if gene_start > start else start)
-
-                splits_dict[split_name] = {'num_genes': len(gene_start_stops),
-                                           'avg_gene_length': numpy.mean([(l[1] - l[0]) for l in gene_start_stops]) if len(gene_start_stops) else 0.0,
-                                           'ratio_coding': total_coding_nts * 1.0 / (stop - start),
-                                           }
-
-        # open connection
-        contigs_db = ContigsDatabase(self.db_path)
-        # push raw entries for splits table
-        db_entries = [tuple([split] + [splits_dict[split][h] for h in t.genes_in_splits_summary_table_structure[1:]]) for split in splits_dict]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?)''' % t.genes_in_splits_summary_table_name, db_entries)
-
-        # push entries for genes in splits table
-        db_entries = [tuple([entry_id] + [genes_in_splits.splits_to_prots[entry_id][h] for h in t.genes_in_splits_table_structure[1:]]) for entry_id in genes_in_splits.splits_to_prots]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?)''' % t.genes_in_splits_table_name, db_entries)
-        # disconnect
-        contigs_db.disconnect()
-
-
-class TablesForHMMHits(Table):
-    def __init__(self, db_path, num_threads_to_use=1, run=run, progress=progress):
-        self.num_threads_to_use = num_threads_to_use
-        self.db_path = db_path
-
-        self.debug = False
-
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, run, progress)
-
-        if not self.genes_are_called:
-            raise ConfigError("It seems the contigs database '%s' was created with '--skip-gene-calling' flag.\
-                                Nothing to do here :/" % (self.db_path))
-
-        self.init_gene_calls_dict()
-
-        if not len(self.gene_calls_dict):
-            raise ConfigError("Tables that should contain gene calls are empty. Which probably means the gene\
-                                caller reported no genes for your contigs.")
-
-        self.set_next_available_id(t.hmm_hits_table_name)
-        self.set_next_available_id(t.hmm_hits_splits_table_name)
-
-
-    def populate_search_tables(self, sources={}):
-        # if we end up generating a temporary file for amino acid sequences:
-        if not len(sources):
-            import anvio.data.hmm
-            sources = anvio.data.hmm.sources
-
-        if not sources:
-            return
-
-        target_files_dict = {}
-
-        tmp_directory_path = filesnpaths.get_temp_directory_path()
-
-        # here we will go through targets and populate target_files_dict based on what we find among them.
-        targets = set([s['target'] for s in list(sources.values())])
-        for target in targets:
-
-            alphabet, context = utils.anvio_hmm_target_term_to_alphabet_and_context(target)
-
-            self.run.info('Target found', '%s:%s' % (alphabet, context))
-
-            class Args: pass
-            args = Args()
-            args.contigs_db = self.db_path
-            contigs_db = ContigsSuperclass(args)
-
-            if context == 'GENE':
-                if alphabet == 'AA':
-                    target_files_dict['AA:GENE'] = os.path.join(tmp_directory_path, 'aa_gene_sequences.fa')
-                    self.export_sequences_table_in_db_into_FASTA_file(t.gene_amino_acid_sequences_table_name, output_file_path=target_files_dict['AA:GENE'])
-                else:
-                    target_files_dict['%s:GENE' % alphabet] = os.path.join(tmp_directory_path, '%s_gene_sequences.fa' % alphabet)
-                    contigs_db.gen_FASTA_file_of_sequences_for_gene_caller_ids(output_file_path=target_files_dict['%s:GENE' % alphabet],
-                                                                               simple_headers=True,
-                                                                               rna_alphabet=True if alphabet=='RNA' else False)
-            elif context == 'CONTIG':
-                if alphabet == 'AA':
-                    raise ConfigError("You are somewhere you shouldn't be. You came here because you thought it would be OK\
-                                       to ask for AA sequences in the CONTIG context. The answer to that is 'no, thanks'. If\
-                                       you think this is dumb, please let us know.")
-                else:
-                    target_files_dict['%s:CONTIG' % alphabet] = os.path.join(tmp_directory_path, '%s_contig_sequences.fa' % alphabet)
-                    utils.export_sequences_from_contigs_db(self.db_path,
-                                                           target_files_dict['%s:CONTIG' % alphabet],
-                                                           rna_alphabet=True if alphabet=='RNA' else False)
-
-        commander = HMMer(target_files_dict, num_threads_to_use=self.num_threads_to_use)
-
-        for source in sources:
-            alphabet, context = utils.anvio_hmm_target_term_to_alphabet_and_context(sources[source]['target'])
-
-            kind_of_search = sources[source]['kind']
-            domain = sources[source]['domain']
-            all_genes_searched_against = sources[source]['genes']
-            hmm_model = sources[source]['model']
-            reference = sources[source]['ref']
-
-            hmm_scan_hits_txt = commander.run_hmmscan(source,
-                                                      alphabet,
-                                                      context,
-                                                      kind_of_search,
-                                                      domain,
-                                                      all_genes_searched_against,
-                                                      hmm_model,
-                                                      reference)
-
-            if not hmm_scan_hits_txt:
-                search_results_dict = {}
-            else:
-                parser = parser_modules['search']['hmmscan'](hmm_scan_hits_txt, alphabet=alphabet, context=context)
-                search_results_dict = parser.get_search_results()
-
-            if not len(search_results_dict):
-                run.info_single("The HMM source '%s' returned 0 hits. SAD (but it's stil OK)." % source, nl_before=1)
-
-
-            if context == 'CONTIG':
-                # we are in trouble here. because our search results dictionary contains no gene calls, but contig
-                # names that contain our hits. on the other hand, the rest of the code outside of this if statement
-                # expects a `search_results_dict` with gene callers id in it. so there are two things we need to do
-                # to do. one is to come up with some new gene calls and add them to the contigs database. so things
-                # will go smoothly downstream. two, we will need to update our `search_results_dict` so it looks
-                # like a a dictionary the rest of the code expects with `gene_callers_id` fields. both of these
-                # steps are going to be taken care of in the following function. magic.
-
-                self.run.warning("Alright! You just called an HMM profile that runs on contigs. Because it is not\
-                                 working with anvi'o gene calls directly, the resulting hits will need to be added\
-                                 as 'new gene calls' into the contigs database. This is a new feature, and if it\
-                                 starts screwing things up for you please let us know. Other than that you're pretty\
-                                 much golden. Carry on.",
-                                 header="Psst. Your fancy HMM profile '%s' speaking" % source,
-                                 lc="green")
-
-                num_hits_before = len(search_results_dict)
-                search_results_dict = utils.get_pruned_HMM_hits_dict(search_results_dict)
-                num_hits_after = len(search_results_dict)
-
-                if num_hits_before != num_hits_after:
-                    self.run.info('Pruned', '%d out of %d hits were removed due to redundancy' % (num_hits_before - num_hits_after, num_hits_before))
-
-                search_results_dict = self.add_new_gene_calls_to_contigs_db_and_update_serach_results_dict(kind_of_search, search_results_dict)
-
-            self.append(source, reference, kind_of_search, domain, all_genes_searched_against, search_results_dict)
-
-        if not self.debug:
-            commander.clean_tmp_dirs()
-            for v in list(target_files_dict.values()):
-                os.remove(v)
-
-
-    def add_new_gene_calls_to_contigs_db_and_update_serach_results_dict(self, source, search_results_dict):
-        """Add new gene calls to the contigs database and update the HMM `search_results_dict`.
-
-           When we are looking for HMM hits in the context of CONTIGS, our hits do not
-           related to the gene calls we already have in a given contigs database. One
-           slution is to add additional gene calls for a given set of HMM hits to keep
-           them in the database."""
-
-        # we will first learn the next available id in the gene callers table
-        contigs_db = ContigsDatabase(self.db_path)
-        next_id = contigs_db.db.get_max_value_in_column('genes_in_contigs', 'gene_callers_id') + 1
-        contigs_db.disconnect()
-
-        additional_gene_calls = {}
-        for e in search_results_dict.values():
-            start = e['start']
-            stop = e['stop']
-
-            if stop > start:
-                direction = 'f'
-            else:
-                direction = 'r'
-                stop, start = start, stop
-
-            partial = 0 if ((stop - start) % 3 == 0) else 1
-
-            # add a new gene call in to the dictionary
-            additional_gene_calls[next_id] = {'contig': e['contig_name'],
-                                              'start': start,
-                                              'stop': stop,
-                                              'direction': direction,
-                                              'partial': partial,
-                                              'source': source,
-                                              'version': 'unknown'
-                                            }
-
-            # update the search results dictionary with gene callers id:
-            e['gene_callers_id'] = next_id
-
-            # update the next available gene callers id:
-            next_id += 1
-
-        # update the contigs db with the gene calls in `additional_gene_calls` dict.
-        gene_calls_table = TablesForGeneCalls(self.db_path, run=terminal.Run(verbose=False))
-        gene_calls_table.use_external_gene_calls_to_populate_genes_in_contigs_table(input_file_path=None,
-                                                                                    gene_calls_dict=additional_gene_calls,
-                                                                                    ignore_internal_stop_codons=True)
-        gene_calls_table.populate_genes_in_splits_tables()
-
-        # refresh the gene calls dict
-        self.init_gene_calls_dict()
-
-        self.run.info('Gene calls added to db', '%d (from source "%s")' % (len(additional_gene_calls), source))
-
-        return search_results_dict
-
-
-    def append(self, source, reference, kind_of_search, domain, all_genes, search_results_dict):
-        # we want to define unique identifiers for each gene first. this information will be used to track genes that will
-        # break into multiple pieces due to arbitrary split boundaries. while doing that, we will add the 'source' info
-        # into the dictionary, so it perfectly matches to the table structure
-
-        for entry_id in search_results_dict:
-            hit = search_results_dict[entry_id]
-
-            gene_call = self.gene_calls_dict[hit['gene_callers_id']]
-
-            hit['gene_unique_identifier'] = hashlib.sha224('_'.join([gene_call['contig'], hit['gene_name'], str(gene_call['start']), str(gene_call['stop'])]).encode('utf-8')).hexdigest()
-            hit['source'] = source
-
-        self.delete_entries_for_key('source', source, [t.hmm_hits_info_table_name, t.hmm_hits_table_name, t.hmm_hits_splits_table_name])
-
-        contigs_db = ContigsDatabase(self.db_path)
-
-        # push information about this search result into serach_info table.
-        db_entries = [source, reference, kind_of_search, domain, ', '.join(all_genes)]
-        contigs_db.db._exec('''INSERT INTO %s VALUES (?,?,?,?,?)''' % t.hmm_hits_info_table_name, db_entries)
-
-        # if our search results were empty, we can return from here.
-        if not len(search_results_dict):
-            contigs_db.disconnect()
-            return
-
-        # then populate serach_data table for each contig.
-        db_entries = []
-        for hit in list(search_results_dict.values()):
-            entry_id = self.next_id(t.hmm_hits_table_name)
-            db_entries.append(tuple([entry_id] + [hit[h] for h in t.hmm_hits_table_structure[1:]]))
-            # tiny hack here: for each hit, we are generating a unique id (`entry_id`), and feeding that information
-            #                 back into the dictionary to pass it to processing of splits, so each split-level
-            #                 entry knows who is their parent.
-            hit['hmm_hit_entry_id'] = entry_id
-
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?,?)''' % t.hmm_hits_table_name, db_entries)
-
-        db_entries = self.process_splits(search_results_dict)
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?)''' % t.hmm_hits_splits_table_name, db_entries)
-
-        contigs_db.disconnect()
-
-
-    def process_splits(self, search_results_dict):
-        hits_per_contig = {}
-        for hit in list(search_results_dict.values()):
-            contig_name = self.gene_calls_dict[hit['gene_callers_id']]['contig']
-
-            if contig_name in hits_per_contig:
-                hits_per_contig[contig_name].append(hit)
-            else:
-                hits_per_contig[contig_name] = [hit]
-
-        db_entries_for_splits = []
-
-        for contig in self.contigs_info:
-            if contig not in hits_per_contig:
-                # no hits for this contig. pity!
-                continue
-
-            for split_name in self.contig_name_to_splits[contig]:
-                split_start = self.splits_info[split_name]['start']
-                split_stop = self.splits_info[split_name]['end']
-
-                # FIXME: this really needs some explanation.
-                for hit in hits_per_contig[contig]:
-                    hit_start = self.gene_calls_dict[hit['gene_callers_id']]['start']
-                    hit_stop = self.gene_calls_dict[hit['gene_callers_id']]['stop']
-
-                    if hit_stop > split_start and hit_start < split_stop:
-                        gene_length = hit_stop - hit_start
-                        # if only a part of the gene is in the split:
-                        start_in_split = (split_start if hit_start < split_start else hit_start) - split_start
-                        stop_in_split = (split_stop if hit_stop > split_stop else hit_stop) - split_start
-                        percentage_in_split = (stop_in_split - start_in_split) * 100.0 / gene_length
-
-                        db_entry = tuple([self.next_id(t.hmm_hits_splits_table_name), hit['hmm_hit_entry_id'], split_name, percentage_in_split, hit['source']])
-                        db_entries_for_splits.append(db_entry)
-
-        return db_entries_for_splits
-
-
-class TablesForStates(Table):
-    def __init__(self, db_path):
-        self.db_path = db_path
-        self.states = {}
-
-        Table.__init__(self, self.db_path, utils.get_required_version_for_db(db_path), run, progress)
-
-        self.init()
-
-
-    def init(self):
-        anvio_db = DBClassFactory().get_db_object(self.db_path)
-        self.states = anvio_db.db.get_table_as_dict(t.states_table_name)
-        anvio_db.disconnect()
-
-
-    def list_states(self):
-        state_names = sorted(list(self.states.keys()))
-
-        self.run.warning('', 'AVAILABLE STATES (%d FOUND)' % (len(self.states)), lc='yellow')
-        for state_name in state_names:
-            self.run.info_single('%s (last modified on %s)' % (state_name, self.states[state_name]['last_modified']),
-                                 nl_after = 1 if state_name == state_names[-1] else 0)
-
-
-    def get_state(self, state_id):
-        if state_id not in self.states:
-            return None
-
-        return self.states[state_id]
-
-
-    def store_state(self, state_id, content, last_modified=None):
-        self.remove_state(state_id)
-
-        last_modified = datetime.datetime.now().strftime("%d.%m.%Y %H:%M:%S") if not last_modified else last_modified
-
-        anvio_db = DBClassFactory().get_db_object(self.db_path)
-        anvio_db.db._exec('''INSERT INTO %s VALUES (?,?,?)''' % t.states_table_name, (state_id, content, last_modified))
-        self.states = anvio_db.db.get_table_as_dict(t.states_table_name)
-
-        anvio_db.disconnect()
-
-
-    def remove_state(self, state_id):
-        self.delete_entries_for_key('name', state_id, [t.states_table_name])
-
-
-class TablesForTaxonomy(Table):
-    """Populate all tables with taxonomy information.
-
-       Essentially takes in a dictionary of genes and taxon calls, populates three tables: taxon_names,
-       splits_taxonomy, and gene_taxonomy when 'create' is called."""
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-        self.run = run
-        self.progress = progress
-
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, self.run, self.progress)
-
-        # this class keeps track of genes that occur in splits, and responsible
-        # for generating the necessary table in the contigs database
-        self.genes_in_splits = GenesInSplits()
-
-
-    def create(self, genes_taxonomy_dict, taxon_names_dict, source='unkown source'):
-        self.source = source
-
-        if not self.genes_are_called:
-            raise ConfigError("Something is wrong. The contigs database says that genes were now called, and here\
-                                you are trying to populate taxonomy tables for genes. No, thanks.")
-
-        self.init_gene_calls_dict()
-
-        self.genes_taxonomy_dict = genes_taxonomy_dict
-        self.taxon_names_dict = taxon_names_dict
-
-        self.sanity_check()
-
-        # oepn connection
-        contigs_db = ContigsDatabase(self.db_path)
-
-        self.splits_info = contigs_db.db.get_table_as_dict(t.splits_info_table_name)
-
-        # test whether there are already genes tables populated
-        taxonomy_source = contigs_db.meta['taxonomy_source']
-        if taxonomy_source:
-            self.run.warning('Previous taxonomy information from "%s" is being replaced with the incoming data\
-                              through "%s".' % (taxonomy_source, self.source))
-            contigs_db.db._exec('''DELETE FROM %s''' % (t.splits_taxonomy_table_name))
-            contigs_db.db._exec('''DELETE FROM %s''' % (t.taxon_names_table_name))
-            contigs_db.db._exec('''DELETE FROM %s''' % (t.genes_taxonomy_table_name))
-
-        # populate taxon mames table
-        self.populate_taxon_names_table()
-
-        # populate genes taxonomy table
-        self.populate_genes_taxonomy_table()
-
-        # compute and push split taxonomy information.
-        self.populate_splits_taxonomy_table()
-
-        # set the source
-        contigs_db.db.remove_meta_key_value_pair('taxonomy_source')
-        contigs_db.db.set_meta_value('taxonomy_source', self.source)
-
-        # disconnect like a pro.
-        contigs_db.disconnect()
-
-
-    def populate_genes_taxonomy_table(self):
-        # open connection
-        contigs_db = ContigsDatabase(self.db_path)
-
-        # push taxonomy data
-        db_entries = [(gene_callers_id, self.genes_taxonomy_dict[gene_callers_id]) for gene_callers_id in self.genes_taxonomy_dict]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?)''' % t.genes_taxonomy_table_name, db_entries)
-
-        # disconnect
-        contigs_db.disconnect()
-
-        self.run.info('Genes taxonomy table', 'Taxonomy stored for %d gene calls' % len(db_entries))
-
-
-    def populate_taxon_names_table(self):
-        # open connection
-        contigs_db = ContigsDatabase(self.db_path)
-
-        db_entries = [tuple([t_name_id] + [self.taxon_names_dict[t_name_id][t_level] for t_level in t.taxon_names_table_structure[1:]]) for t_name_id in self.taxon_names_dict]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?,?)''' % t.taxon_names_table_name, db_entries)
-
-        contigs_db.disconnect()
-        self.run.info('Taxon names table', 'Updated with %d unique taxon names' % len(db_entries))
-
-
-    def sanity_check(self):
-        """Basic checks to make sure things are in order at least to a minimum extent."""
-
-        self.progress.new('Sanity checking ...')
-        self.progress.update('Comparing gene caller ids in the incoming data and in the contigs database ..')
-        gene_caller_ids_in_database = set(self.gene_calls_dict.keys())
-        gene_caller_ids_in_taxonomy_dict = set(self.genes_taxonomy_dict.keys())
-        gene_caller_ids_missing_in_db = gene_caller_ids_in_taxonomy_dict.difference(gene_caller_ids_in_database)
-        self.progress.end()
-
-        run.info("Num gene caller ids in the db", pp(len(gene_caller_ids_in_database)))
-        run.info("Num gene caller ids in the incoming data", pp(len(gene_caller_ids_in_taxonomy_dict)))
-
-        if gene_caller_ids_missing_in_db:
-            raise ConfigError("Taxonomy information for genes you are trying to import into the database contains\
-                                %s gene caller ids that do not appear to be in the database. This is a step you must\
-                                be very careful to make sure you are not importing annotations for genes that have\
-                                nothing to do with your contigs database. To make sure of that, you should always work\
-                                with `anvi-get-dna-sequences-for-gene-calls` or `anvi-get-aa-sequences-for-gene-calls` programs\
-                                to get the data to annotate. For instance one of the gene caller ids you have in your\
-                                input data that does not appear in the database is this one: '%s'. Anvi'o hopes it makes\
-                                sense to you, because it definitely does not make any sense to anvi'o :("\
-                                                        % (len(gene_caller_ids_missing_in_db), str(gene_caller_ids_missing_in_db.pop())))
-
-        # check whether input matrix dict
-        keys_found =  list(self.taxon_names_dict.values())[0].keys()
-        missing_keys = [key for key in t.taxon_names_table_structure[1:] if key not in keys_found]
-        if len(missing_keys):
-            raise ConfigError("Anvi'o is trying to get ready to create tables for taxonomy, but there is something\
-                                wrong :( The taxonomy names dict (one of the required input dictionaries to the class\
-                                seems to be missing a one or more keys that are necessary to finish the job. Here is \
-                                a list of missing keys: %s. The complete list of input keys should contain these: %s."\
-                                        % (', '.join(missing_keys), ', '.join(t.taxon_names_table_structure[1:])))
-
-        if not len(self.taxon_names_dict):
-            raise ConfigError("Anvi'o is trying to get ready to create tables for taxonomy, but taxonomy names dict\
-                                (one of the required input dictionaries to the class responsible for this task) seems\
-                                to be empty.")
-
-
-    def populate_splits_taxonomy_table(self):
-        """Populate the taxonomy information per split"""
-
-        # build a dictionary for fast access to all genes identified within a contig
-        gene_caller_ids_in_contigs = {}
-        for gene_callers_id in self.genes_taxonomy_dict:
-            contig = self.gene_calls_dict[gene_callers_id]['contig']
-            if contig in gene_caller_ids_in_contigs:
-                gene_caller_ids_in_contigs[contig].add(gene_callers_id)
-            else:
-                gene_caller_ids_in_contigs[contig] = set([gene_callers_id])
-
-        contigs_without_annotation = list(set(self.contigs_info.keys()) - set(gene_caller_ids_in_contigs.keys()))
-
-        for contig in contigs_without_annotation:
-            gene_caller_ids_in_contigs[contig] = set([])
-
-        splits_dict = {}
-
-        num_splits_processed = 0
-        num_splits_with_taxonomy = 0
-
-        for contig in self.contigs_info:
-            for split_name in self.contig_name_to_splits[contig]:
-                num_splits_processed += 1
-
-                splits_dict[split_name] = None
-                start = self.splits_info[split_name]['start']
-                stop = self.splits_info[split_name]['end']
-
-                taxon_name_ids = []
-                for gene_callers_id in gene_caller_ids_in_contigs[contig]:
-                    if self.gene_calls_dict[gene_callers_id]['stop'] > start and self.gene_calls_dict[gene_callers_id]['start'] < stop:
-                        taxon_name_ids.append(self.genes_taxonomy_dict[gene_callers_id])
-
-                if not taxon_name_ids:
-                    continue
-
-                if len(set(taxon_name_ids)) == 1:
-                    splits_dict[split_name] = taxon_name_ids[0]
-                else:
-                    d = Counter()
-                    for taxon_name_id in taxon_name_ids:
-                        d[taxon_name_id] += 1
-
-                    most_frequent_taxon_name_id, occurrence = d.most_common()[0]
-                    splits_dict[split_name] = most_frequent_taxon_name_id
-
-                num_splits_with_taxonomy += 1
-
-        # open connection
-        contigs_db = ContigsDatabase(self.db_path)
-
-        # push taxonomy data
-        db_entries = [(split, splits_dict[split]) for split in splits_dict]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?)''' % t.splits_taxonomy_table_name, db_entries)
-
-        # disconnect
-        contigs_db.disconnect()
-
-        self.run.info('Splits taxonomy', 'Input data from "%s" annotated %d of %d splits (%.1f%%) with taxonomy.'\
-                                            % (self.source, num_splits_with_taxonomy, num_splits_processed,
-                                               num_splits_with_taxonomy * 100.0 / num_splits_processed))
-
-
-class TableForGeneClusters(Table):
-    """A class to populte gene clusters table in a given pan db.
-
-      Here is an example:
-
-        >>> table = TableForGeneClusters(db_path)
-        >>> for ...:
-        >>>     table.add({'gene_caller_id': gene_caller_id, 'gene_cluster_id': gene_cluster_id, 'genome_name': genome_name})
-        >>> table.store()
-    """
-
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-
-        is_pan_db(db_path)
-
-        self.run = run
-        self.progress = progress
-
-        Table.__init__(self, self.db_path, anvio.__pan__version__, run, progress)
-
-        self.set_next_available_id(t.pan_gene_clusters_table_name)
-
-        self.entries = []
-
-
-    def add(self, entry_dict):
-        self.entries.append([entry_dict[key] for key in t.pan_gene_clusters_table_structure[1:]])
-
-
-    def store(self):
-        self.delete_contents_of_table(t.pan_gene_clusters_table_name, warning=False)
-
-        db_entries = [tuple([self.next_id(t.pan_gene_clusters_table_name)] + entry) for entry in self.entries]
-        pan_db = PanDatabase(self.db_path)
-        pan_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?)''' % t.pan_gene_clusters_table_name, db_entries)
-        pan_db.disconnect()
-
-
-class TableForGeneFunctions(Table):
-    def __init__(self, db_path, run=run, progress=progress):
-        self.db_path = db_path
-
-        self.run = run
-        self.progress = progress
-
-        Table.__init__(self, self.db_path, anvio.__contigs__version__, run, progress)
-
-        self.set_next_available_id(t.gene_function_calls_table_name)
-
-
-    def create(self, functions_dict, drop_previous_annotations_first = False):
-        self.sanity_check()
-
-        # incoming stuff:
-        gene_function_sources = set([v['source'] for v in list(functions_dict.values())])
-        unique_num_genes = len(set([v['gene_callers_id'] for v in list(functions_dict.values())]))
-
-        # oepn connection
-        contigs_db = ContigsDatabase(self.db_path)
-
-        # are there any previous annotations in the db:
-        gene_function_sources_in_db = set(contigs_db.meta['gene_function_sources'] or [])
-
-        # difference between sources in the db, and incoming sources:
-        gene_function_sources_both_in_db_and_incoming_dict = gene_function_sources.intersection(gene_function_sources_in_db)
-
-        # here we will do some magic. there are mulitple scenarios to consider here based on whether there
-        # are functions already in the database, whether some of them matches to the incoming functions, etc.
-        # let's go case-by-case:
-        if not gene_function_sources_in_db:
-            # set the sources and continue
-            contigs_db.db.remove_meta_key_value_pair('gene_function_sources')
-            contigs_db.db.set_meta_value('gene_function_sources', ','.join(list(gene_function_sources)))
-
-        elif gene_function_sources_in_db and drop_previous_annotations_first:
-            # there are gene calls, but the user wants everything to be dropeped.
-            self.run.warning("As per your request, anvi'o is DROPPING all previous function calls from %d sources\
-                              before adding the incoming data, which contains %d entries originating from %d sources: %s" \
-                                    % (len(gene_function_sources_in_db), len(functions_dict),
-                                       len(gene_function_sources), ', '.join(gene_function_sources)))
-
-            # clean the table and reset the next available ids
-            contigs_db.db._exec('''DELETE FROM %s''' % (t.gene_function_calls_table_name))
-            self.reset_next_available_id_for_table(t.gene_function_calls_table_name)
-
-            # set the sources
-            contigs_db.db.remove_meta_key_value_pair('gene_function_sources')
-            contigs_db.db.set_meta_value('gene_function_sources', ','.join(gene_function_sources))
-
-        elif gene_function_sources_in_db and gene_function_sources_both_in_db_and_incoming_dict:
-            # some of the functions in the incoming dict match to what is already in the db. remove
-            self.run.warning("Some of the annotaiton sources you want to add into the database are already in the db. So\
-                              anvi'o will REPLACE those with the incoming data from these sources: %s" % \
-                                            ', '.join(gene_function_sources_both_in_db_and_incoming_dict))
-
-            # remove those entries for matching sources:
-            for source in gene_function_sources_both_in_db_and_incoming_dict:
-                contigs_db.db._exec('''DELETE FROM %s WHERE source = "%s"''' % (t.gene_function_calls_table_name, source))
-
-            # set the sources
-            contigs_db.db.remove_meta_key_value_pair('gene_function_sources')
-            contigs_db.db.set_meta_value('gene_function_sources', ','.join(list(gene_function_sources_in_db.union(gene_function_sources))))
-
-        else:
-            # fuctions in the db, but none of them match with the incoming annotation sources. totally new stuff.
-            # good then. update sources
-            contigs_db.db.remove_meta_key_value_pair('gene_function_sources')
-            contigs_db.db.set_meta_value('gene_function_sources', ','.join(list(gene_function_sources_in_db.union(gene_function_sources))))
-
-        # push the data
-        db_entries = [tuple([self.next_id(t.gene_function_calls_table_name)] + [functions_dict[v][h] for h in t.gene_function_calls_table_structure[1:]]) for v in functions_dict]
-        contigs_db.db._exec_many('''INSERT INTO %s VALUES (?,?,?,?,?,?)''' % t.gene_function_calls_table_name, db_entries)
-
-        # disconnect like a pro.
-        contigs_db.disconnect()
-
-        self.run.info('Gene functions', '%d function calls from %d sources for %d unique gene calls has\
-                                        been added to the contigs database.' % \
-                                            (len(functions_dict), len(gene_function_sources), unique_num_genes))
-
-
-    def sanity_check(self):
-        pass
-
-
-class TableForNtPositions(object):
-    def __init__(self, run=run, progress=progress):
-        self.run = run
-        self.progress = progress
-        self.numpy_data_type = 'uint8'
-        self.db_entries = []
-
-
-    def append(self, contig_name, position_info_list):
-        position_info_blob = utils.convert_numpy_array_to_binary_blob(numpy.array(position_info_list, dtype=self.numpy_data_type))
-        self.db_entries.append((contig_name, position_info_blob, ))
-
-
-    def store(self, db):
-        db.insert_many(t.nt_position_info_table_name, entries=self.db_entries)
-        self.db_entries = []
-
-
-class GenesInSplits:
-    def __init__(self):
-        self.entry_id = 0
-        self.splits_to_prots = {}
-
-    def add(self, split_name, split_start, split_end, gene_callers_id, prot_start, prot_end):
-
-        gene_length = prot_end - prot_start
-
-        if gene_length <= 0:
-            raise ConfigError("dbops.py/GeneInSplits: OK. There is something wrong. We have this gene, '%s',\
-                                which starts at position %d and ends at position %d. Well, it doesn't look right,\
-                                does it?" % (gene_callers_id, prot_start, prot_end))
-
-        # if only a part of the gene is in the split:
-        start_in_split = (split_start if prot_start < split_start else prot_start) - split_start
-        stop_in_split = (split_end if prot_end > split_end else prot_end) - split_start
-        percentage_in_split = (stop_in_split - start_in_split) * 100.0 / gene_length
-
-        self.splits_to_prots[self.entry_id] = {'split': split_name,
-                                               'gene_callers_id': gene_callers_id,
-                                               'start_in_split': start_in_split,
-                                               'stop_in_split': stop_in_split,
-                                               'percentage_in_split': percentage_in_split}
-        self.entry_id += 1
-
-
 ####################################################################################################
 #
 #     HELPER FUNCTIONS
 #
 ####################################################################################################
-
-def is_contigs_db(db_path):
-    filesnpaths.is_file_exists(db_path)
-    if utils.get_db_type(db_path) != 'contigs':
-        raise ConfigError("'%s' is not an anvi'o contigs database." % db_path)
-
-
-def is_pan_or_profile_db(db_path):
-    if utils.get_db_type(db_path) not in ['pan', 'profile']:
-        raise ConfigError("'%s' is neither a pan nor a profile database :/ Someone is in trouble." % db_path)
-
-
-def is_profile_db(db_path):
-    if utils.get_db_type(db_path) != 'profile':
-        raise ConfigError("'%s' is not an anvi'o profile database." % db_path)
-
-
-def is_pan_db(db_path):
-    if utils.get_db_type(db_path) != 'pan':
-        raise ConfigError("'%s' is not an anvi'o pan database." % db_path)
-
 
 def is_db_ok_to_create(db_path, db_type):
     if os.path.exists(db_path):
@@ -4713,8 +2856,8 @@ def is_blank_profile(db_path):
 
 
 def is_profile_db_and_contigs_db_compatible(profile_db_path, contigs_db_path):
-    is_contigs_db(contigs_db_path)
-    is_profile_db(profile_db_path)
+    utils.is_contigs_db(contigs_db_path)
+    utils.is_profile_db(profile_db_path)
 
     contigs_db = ContigsDatabase(contigs_db_path)
     profile_db = ProfileDatabase(profile_db_path)
@@ -4737,7 +2880,7 @@ def is_profile_db_and_contigs_db_compatible(profile_db_path, contigs_db_path):
 
 
 def get_split_names_in_profile_db(profile_db_path):
-    is_profile_db(profile_db_path)
+    utils.is_profile_db(profile_db_path)
 
     profile_db = ProfileDatabase(profile_db_path)
