+++ anvio/tables/contigsplitinfo.py
@@ -1,0 +1,72 @@
+# -*- coding: utf-8
+# pylint: disable=line-too-long
+
+
+import anvio
+import anvio.tables as t
+import anvio.utils as utils
+import anvio.terminal as terminal
+
+
+__author__ = "Developers of anvi'o (see AUTHORS.txt)"
+__copyright__ = "Copyleft 2015-2018, the Meren Lab (http://merenlab.org/)"
+__credits__ = []
+__license__ = "GPL 3.0"
+__version__ = anvio.__version__
+__maintainer__ = "A. Murat Eren"
+__email__ = "a.murat.eren@gmail.com"
+__status__ = "Development"
+
+
+run = terminal.Run()
+progress = terminal.Progress()
+pp = terminal.pretty_print
+
+
+class TableForContigsInfo:
+    def __init__(self, split_length):
+        self.db_entries = []
+        self.total_nts = 0
+        self.total_contigs = 0
+        self.split_length = split_length
+
+
+    def append(self, seq_id, sequence, gene_start_stops=None):
+        sequence_length = len(sequence)
+        gc_content = utils.get_GC_content_for_sequence(sequence)
+
+        # how many splits will there be?
+        split_start_stops = utils.get_split_start_stops(sequence_length, self.split_length, gene_start_stops)
+
+        self.total_nts += sequence_length
+        self.total_contigs += 1
+        db_entry = tuple([seq_id, sequence_length, gc_content, len(split_start_stops)])
+        self.db_entries.append(db_entry)
+
+        return (sequence_length, split_start_stops, gc_content)
+
+
+    def store(self, db):
+        if len(self.db_entries):
+            db._exec_many('''INSERT INTO %s VALUES (%s)''' % (t.contigs_info_table_name, (','.join(['?'] * len(self.db_entries[0])))), self.db_entries)
+
+
+class TableForSplitsInfo:
+    def __init__(self):
+        self.db_entries = []
+        self.total_splits = 0
+
+
+    def append(self, seq_id, sequence, order, start, end, parent_gc_content, parent):
+        self.total_splits += 1
+        sequence_length = len(sequence)
+        db_entry = tuple([seq_id, order, start, end, sequence_length, utils.get_GC_content_for_sequence(sequence), parent_gc_content, parent])
+        self.db_entries.append(db_entry)
+
+
+    def store(self, db):
+        if len(self.db_entries):
+            db._exec_many('''INSERT INTO %s VALUES (%s)''' % (t.splits_info_table_name, (','.join(['?'] * len(self.db_entries[0])))), self.db_entries)
+
+
+
